backend/requirements.txt
fastapi
uvicorn
pydantic
requests
python-multipart
python-dotenv
sentence-transformers
faiss-cpu
beautifulsoup4
pymupdf
python-docx
langchain-text-splitters
